phase-4_integration/requirements.txt
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
google-genai>=1.0
python-dotenv>=1.0
streamlit>=1.30
